requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distlib==0.3.8
Django==4.2.14
django-celery-beat==2.6.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
environ==1.0
filelock==3.15.4
h11==0.14.0
idna==3.7
kombu==5.3.7
platformdirs==4.2.2
prompt-toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
standard==1.0.3
termcolor==2.4.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn[standard]
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
